Splash screen and Mux
- Added a way to acces the 8 to 1 multiplexer
- splash screen with compile time and date

+++ drumkit.h
@@ -396,7 +396,7 @@ class DrumkitClass
 	{
 		sampleIndex[index][1] = gain;
 	}
-	void setPan(uint16_t index, int16_t pan)
+	void setPan(uint16_t index, int8_t pan)
 	{
 		if(pan < -127) pan = -127;
 		if(pan > 127) pan = 127;
